% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squirrel.R
\name{squirrel}
\alias{squirrel}
\title{Squirrel}
\usage{
squirrel(site = NULL)
}
\arguments{
\item{site}{character, URL of the site you would like to be sent to
as a welcomed distraction. If `NULL`, will default to randomly selecting
from one of the curated sites above.}
}
\description{
For facilitating a lovely distraction. Sends the user randomly to
* https://twitter.com
* https://xkcd.com
* http://phdcomics.com
* https://www.reddit.com
or a site of your choice if the `site` argument is populated.
}
